split up into different files, re-enable other commands
